Extract macros to Item data class & Extract composable funcs

+++ app/src/main/java/com/example/nutrition_tracker/data/Macro.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
-data class Macro(val name: String, val amount: Double, val color: Color,
+data class Macro(val name: String, val amount: Int, val color: Color,
                  val startPadding: Dp = 0.dp,
                  val endPadding: Dp = 0.dp,
                  val topLeft: Dp = 0.dp, val topRight: Dp = 0.dp,
